+++ test/store/index.js
@@ -429,37 +429,19 @@ describe('store', () => {
 			assert.equal(b_started, false);
 		});
 
-		it('works with undefined stores #1', () => {
-			const a = derived(null, (n) => {
-				return n;
+		it('errors on undefined stores #1', () => {
+			assert.throws(() => {
+				derived(null, (n) => n);
 			});
-			const values = [];
-			const unsubscribe = a.subscribe((value) => values.push(value));
-			unsubscribe();
-			assert.deepEqual(values, [undefined]);
 		});
 
-		it('works with undefined stores #2', () => {
-			const a = writable(1);
-			const b = derived([a, null, undefined], ([n, un1, un2]) => {
-				assert.equal(un1, undefined);
-				assert.equal(un2, undefined);
-				return n * 2;
+		it('errors on undefined stores #2', () => {
+			assert.throws(() => {
+				const a = writable(1);
+				derived([a, null, undefined], ([n]) => {
+					return n * 2;
+				});
 			});
-
-			const values = [];
-
-			const unsubscribe = b.subscribe(value => {
-				values.push(value);
-			});
-
-			a.set(2);
-			assert.deepEqual(values, [2, 4]);
-
-			unsubscribe();
-
-			a.set(3);
-			assert.deepEqual(values, [2, 4]);
 		});
 	});